higeco-signing-keys: create recipe for custom packaging keys delivery (#26)
Recipe to deliver a series of GPG keys for RPM (and in principle OPKG) packaging.

It should be very simple, just drop the keys inside the higeco-signing-keys/keys subdirectory (there is an example one, the one from the distro) and it will be created. Indications on how to manage the versioning (for upgrades) is in the recipe.

If this looks good after merging please proceed with adding keys and so on. A reference to this package (specifically higeco-signing-keys-rpm for the RPM based image) should be then added also to the image recipe to have it there. The example key can be of course removed (I've put it there for testing, with nothing would of course complain)